% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ageretr}
\alias{ageretr}
\title{Une table avec des séries chronologiques d'âges moyens de départ à la retraite}
\format{
Un data frame avec 15932 observations et 8 variables:
\describe{
\item{cc}{Code caisse, selon la nomenclature utilisée par la DREES pour l'EACR en open data}
\item{caisse}{Intitulé de la caisse de retraite}
\item{source}{Source pour chaque valeur : EACR ou 'chaînage historique'}
\item{sexe}{Sexe}
\item{naiss}{Lieu de naissance : France, Etranger ou Ensemble}
\item{generation}{Année de naissance}
\item{champ_obs}{Champ des retraités observés à 67 ou 70 ans}
\item{ageliq}{Âge moyen de liquidation sur le champ considéré}
}
}
\usage{
ageretr
}
\description{
La table contient des séries chronologiques d'âge moyen d'entrée en jouissance d'un droit direct de retraite
par génération, sexe, régime, lieu de naissance. Les données sont extraites de l'EACR.
}
\details{
Les âges moyens sont calculés pour les retraités en vie et ayant déjà liquidé un droit à un âge donné (67 ou 70 ans).
Pour les générations qui ne sont pas observées à ces âges-là dans l'EACR, on a chaîné les évolutions de génération en génération à âge donné,
de façon à la recaler sur la génération la plus ancienne observée à l'âge considéré.

Le recul historique dans l'EACR est plus ou moins long selon les régimes. Dans certains cas,
il y a manifestement des ruptures de séries certaines années, induisant un "saut" dans les âges moyens de
départ à la retraite par génération. Lorsque de tels sauts ont été identifiés, on a tronqué les données pour
ne retenir que les années d'observations les plus récentes (afin d'avoir des observations homogènes
sans rupture de série).
}
\keyword{datasets}
