% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{docs_cor}
\alias{docs_cor}
\title{Table des documents diffusés dans le cadre des séances mensuelles du COR}
\format{
A data frame with 2332 observations and 19 variables:
\describe{
\item{seance}{Titre de la séance}
\item{date.seance}{Date de la séance}
\item{url.seance}{url de la page présentant la séance sur le site internet du COR}
\item{annee}{Année de la séance}
\item{mois}{Mois de la séance}
\item{jour}{Jour de la séance}
\item{texte.introductif}{Texte de présentation de la séance}
\item{mots.cles}{Mots clés de la séance (tels que fournis par le COR)}
\item{partie}{Titre de la partie}
\item{titre_complet}{Titre complet du document}
\item{auteur}{Auteur(s) du document}
\item{url.doc}{Url du document sur le site internet du COR}
\item{type}{Type de document (document, diaporama, fichier excel...)}
\item{titre}{Titre du document}
\item{num_document}{Numéro du document}
\item{seance_dapres_pdf}{Titre de la séance, lu dans le fichier pdf}
\item{numdocument_dapres_pdf}{Numero du document, lu dans le fichier pdf}
\item{titre_dapres_pdf}{Titre du document, lu dans le fichier pdf}
\item{auteur_dapres_pdf}{Auteur(s) du document, lu dans le fichier pdf}
}
}
\usage{
docs_cor
}
\description{
(version provisoire, encore en cours de construction)
}
\details{
Données extraites fin août 2022, s'arrêtant à la séance de juillet 2022.
}
\keyword{datasets}
