% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dossiers_en_bref_cor}
\alias{dossiers_en_bref_cor}
\title{Une base des questions/réponses extraites des documents "Le dossier en bref" du COR}
\format{
A data frame with 582 observations and 7 variables:
\describe{
\item{question}{Texte de la question extrait du document}
\item{reponse}{Réponse à la question extrait du document}
\item{reponse_avec_url}{Réponse à la question extrait du document, avec lien hypertexte vers les documents cités}
\item{partie}{Titre intermédiaire extrait du document}
\item{url.doc}{url du document sur le site internet du COR}
\item{date.seance}{Date de la séance}
\item{titre.seance}{Titre de la séance}
}
}
\usage{
dossiers_en_bref_cor
}
\description{
(version provisoire, encore en cours de construction)
}
\details{
Données extraites fin août 2022, s'arrêtant à la séance de juillet 2022.
}
\keyword{datasets}
