% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrait_eacr.R
\name{extrait_eacr}
\alias{extrait_eacr}
\title{Extrait les données opendata de l'enquête annuelle auprès des caisses de retraite}
\usage{
extrait_eacr(
 
    fichier = c("https://data.drees.solidarites-sante.gouv.fr/api/datasets/1.0/donnes_eacr/attachments/eacr_diffusee_part_2_version_du_20_septembre_2022_xlsx/",
    "https://data.drees.solidarites-sante.gouv.fr/api/datasets/1.0/donnes_eacr/attachments/eacr_diffusee_part_1_version_du_20_septembre_2022_xlsx/"),
  tableaux = NULL
)
}
\arguments{
\item{fichier}{emplacement du fichier Excel contenant les données}

\item{tableaux}{nom des tableaux (onglets du fichier Excel)}
}
\value{
une liste de data.frame
}
\description{
Si l'adresse du fichier n'est pas indiquée, la fonction lit par défaut la dernière version des données disponibles sur l'espace data.Drees (version diffusée le 10 octobre 2022).
}
