% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrait_tableau_ccss.R
\name{extrait_tableau_ccss}
\alias{extrait_tableau_ccss}
\title{Extrait des informations d'un tableau du rapport de la CCSS sous forme d'une table de données}
\usage{
extrait_tableau_ccss(pdf_ccss = NULL, page = 1, position = 1)
}
\arguments{
\item{pdf_ccss}{emplacement du rapport CCSS (fichier PDF)}

\item{page}{page du rapport sur laquelle se situe le tableau qu'on souhaite extraire}

\item{position}{position (dans l'ordre : 1, 2 ...) du tableau sur la page}
}
\value{
un dataframe
}
\description{
Lit une page donnée d'un rapport de la CCSS, et en extrait les données chiffrées issues d'un tableau,
pour les restituer sous la forme d'un data.frame
}
