% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_modalites_eir.R
\name{recode_modalites_eir}
\alias{recode_modalites_eir}
\title{Recode une variable de l'EIR en remplaçant ses modalités par leurs intitulés}
\usage{
recode_modalites_eir(var, nomvar)
}
\arguments{
\item{var}{la variable à recoder}

\item{nomvar}{le nom de la variable parmi les noms des variables de l'EIR (character)}
}
\value{
une variable dont les modalités sont des intitulés en clair
}
\description{
Recode une variable de l'EIR en remplaçant ses modalités par leurs intitulés
}
