% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{revalopensions}
\alias{revalopensions}
\title{Une table des indices de revalorisations des pensions dans les principaux régimes}
\format{
Un data frame avec 4865 observations et 5 variables:
\describe{
\item{cc}{Code caisse, selon la nomenclature utilisée par la DREES pour l'EIR et l'EACR en open data}
\item{date}{Date (toujours au 1er de chaque mois)}
\item{annee}{Année}
\item{mois}{Mois}
\item{indicerevalo}{Indice cumulé de revalorisation à la date indiquée}
}
}
\usage{
revalopensions
}
\description{
La table contient des indices de revalorisations cumulés : pour obtenir les coefficients de revalorisation entre
deux dates, il faut donc faire le ratio des indices entre ces deux dates.
}
\details{
Les régimes sont repérés par leur code "CC", cohérent avec les données de l'EACR diffusées par la DREES en open data.
Les données sont issues des barèmes IPP : https://www.ipp.eu/baremes-ipp/regimes-de-retraites/
}
\keyword{datasets}
