% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seances_cor}
\alias{seances_cor}
\title{Index des titres et dates des séances mensuelles du COR}
\format{
A data frame with 304 observations and 8 variables:
\describe{
\item{titre.seance}{Titre de la séance}
\item{date.seance}{Date de la séance}
\item{url.seance}{url de la page présentant la séance sur le site internet du COR}
\item{annee}{Année de la séance}
\item{mois}{Mois de la séance}
\item{jour}{Jour de la séance}
\item{texte.introductif}{Texte de présentation de la séance}
\item{mots.cles}{Mots clés de la séance (tels que fournis par le COR)}
}
}
\usage{
seances_cor
}
\description{
(version provisoire, encore en cours de construction)
}
\details{
Données extraites fin août 2022, s'arrêtant à la séance de juillet 2022.
}
\keyword{datasets}
