% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{txretr}
\alias{txretr}
\title{Taux de retraités par âge fin entre 50 et 70 ans (source DREES, modèle ANCETRE)}
\format{
A data frame with 1134 observations and 7 variables:
\describe{
\item{age}{Age, entre 50 et 70 ans}
\item{annee}{Année}
\item{txretr}{Taux de retraités, en \% (chiffre en 0 et 1)}
\item{txnouvretr}{Taux de nouveaux retraités au cours de l'année, en \% (chiffre en 0 et 1)}
\item{sexe}{ensemble, femmes, hommes}
\item{geo}{lieu de résidence}
\item{datepubli}{édition du Panorama 'retraités et retraites' dont les données sont tirées}
}
}
\usage{
txretr
}
\description{
Les données proviennent du panorama annuel de la DREES "retraités et retraites", à partir du modèle ANCETRE.
}
\keyword{datasets}
