% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variables_eir}
\alias{variables_eir}
\title{Une table des variables de l'EIR et de leurs modalités}
\format{
A data frame with 582 observations and 7 variables:
\describe{
\item{variable}{Nom de la variable}
\item{libelle}{Libellé de la variable}
\item{modalites}{Liste des modalités de la variable (séparées par un ;)}
\item{table}{Table dans laquelle la variable est disponible (avant, ddir, etc.)}
}
}
\usage{
variables_eir
}
\description{
Données récupérées sur le site du CASD.
}
\details{
Voir aussi la table 'modalites_eir'.
}
\keyword{datasets}
